% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_filter.R
\name{data_filter}
\alias{data_filter}
\title{Keep rows that match a condition}
\usage{
data_filter(data, dic, var_inputs, special_placeholder = NULL, .id = NULL)
}
\arguments{
\item{data}{A data frame}

\item{dic}{A data frame dictionary. This can be created by \code{create_dic()} function from \code{homodatum}package. Creating a dictionary allow \code{data_filter()} to detect different types of values (Dat, Cat, Num or list). This types of values belong to homodatum schema and can be checked by \code{homodatum::available_hd_Types()} function.}

\item{var_inputs}{A list where columns columns and condition must be specified.}

\item{special_placeholder}{Default as NULL. A character ("Todos" / "All") that adds a clickable filtering option when building a shiny app.}

\item{.id}{a character that specifies a the data frame variable id.}
}
\description{
The \code{data_filter()} function is used to subset a data frame, retaining all rows that satisfy a given condition.
}
\examples{

data <- iris
dic <- homodatum::create_dic(data)
names(data) <- dic$id
var_inputs <- list("species" = c("All"))
data_result <- data_filter(data,
                           dic,
                           var_inputs = var_inputs,
                           special_placeholder = "All")

 # checking results
 unique(data_result$species) == unique(data$species)

}
