% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_var_aggregation.R
\name{var_aggregation}
\alias{var_aggregation}
\title{Compute Summary Statistics of Data}
\usage{
var_aggregation(data, dic = NULL, agg, group_var, to_agg, name = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{dic}{A data frame dictionary. This can be created by \code{create_dic()} function from \code{homodatum}package. Creating a dictionary allow \code{data_filter()} to detect different types of values (Dat, Cat, Num or list). This types of values belong to homodatum schema and can be checked by \code{homodatum::available_hd_Types()} function.}

\item{agg}{Type of aggregation to be calculated. If it exists in base R, it can be calculated (sum, count, mean, sd, etc.).}

\item{group_var}{variable(s) to group by.}

\item{to_agg}{variable to be aggregated when more than one are available.}

\item{name}{The name of the new column in the output.}
}
\description{
The \code{var_aggregation()} function is used to compute custom summary statistics from a data frame. This function helps reducing code writing and optimization when building a shiny app.
}
\examples{

data <- homodatum::sample_data("Cat-Cat-Num-Num-Cat")

names(data) <- c("a", "b", "c", "d", "e")

dic <- homodatum::create_dic(data)

names(data) <- dic$id

data_result <- var_aggregation(data = data,
                               dic = dic,
                               agg = "count",
                               to_agg = "c",
                               name = "Conteo",
                               group_var = c("a", "b"))

}
