% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_variable_selection.R
\name{variable_selection}
\alias{variable_selection}
\title{Subset columns by their names}
\usage{
variable_selection(data, path = NULL, viz = NULL, ...)
}
\arguments{
\item{data}{A data frame}

\item{path}{path to the YAML to be called if necessary. Default value is NULL.}

\item{viz}{Specific type of viz when selecting by a YAML configuration file. Default value is NULL.}

\item{...}{One or more quoted expressions (variable names) separated by commas.}
}
\description{
The \code{data_select()} function is used to subset an specific columns from a data frame by their names or by defining them in a YAML configuration file. This function helps reducing code writing when building a shiny app.
}
\examples{

data <- iris

data_result <- variable_selection(data = iris,
                                  viz = NULL,
                                  path = NULL,
                                  "Sepal.Length", "Petal.Length")

 names(data_result)

}
