% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_viz_selection.R
\name{viz_selection}
\alias{viz_selection}
\title{Choose the proper visualization for you data}
\usage{
viz_selection(data, dic, viz, num_hType = FALSE)
}
\arguments{
\item{data}{A data frame to use for plot}

\item{dic}{A data frame dictionary. This can be created by \code{create_dic()} function from \code{homodatum}package. Creating a dictionary allow \code{data_filter()} to detect different types of values (Dat, Cat, Num or list). This types of values belong to homodatum schema and can be checked by \code{homodatum::available_hd_Types()} function.}

\item{viz}{Type of visualization needed. Options available: \emph{map, map_bubbles, line, bar}.}

\item{num_hType}{helper to set if the column with the data to be plotted is a Num type, from the \code{available_hdTypes()}.}
}
\value{
type of visualization needed from the \code{hgchmagic} package
}
\description{
The \code{var_aggregation()} function is used to compute custom summary statistics from a data frame. This function helps reducing code writing and optimization when building a shiny app.
}
\examples{

data <- iris
dic <- homodatum::create_dic(data)
names(data) <- dic$id
data_viz <- variable_selection(data = data, path = NULL, viz = NULL, "species")
data_viz  <- var_aggregation(data_viz,
                             dic, agg = "count",
                             group_var = "species",
                             to_agg = "species",
                             name = "Total")
viz_result <- viz_selection(data = data_viz,
                             dic = dic,
                             viz = "map", num_hType = TRUE)

}
