% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_write_html.R
\name{write_html_group}
\alias{write_html_group}
\title{Write HTML output for grouped data}
\usage{
write_html_group(
  data,
  dic,
  click,
  text_result_null = NULL,
  separate_row = NULL,
  sep_url = ",",
  url_name = "Link &#8734;",
  url_class = "url-class",
  id = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame to create the output from}

\item{dic}{A list of character vectors where each vector contains the names of the variables to be included in the output. The names of the list should be the group names}

\item{click}{A vector of variable names from data that will be used to filter the data. The HTML output is created only from the filtered data}

\item{text_result_null}{Text to be returned in case the filtered data is empty}

\item{separate_row}{Variable name that should be used to separate the rows in case it has multiple values}

\item{sep_url}{Separator used to separate multiple URLs in case they exist}

\item{url_name}{Name used to show the URL link in the HTML output}

\item{url_class}{CSS class for the URL link}

\item{id}{A variable name to use as an ID for the HTML output}

\item{...}{Other arguments to be passed to \code{dplyr::select()}}
}
\value{
An HTML output for the filtered data
}
\description{
This function creates an HTML output for a grouped data set.
}
